==> code/03_Figure3_dynamics.Rmd <==
---
title: 'Figure 3: Comparing dynamic responses'
output: github_document
---
This notebook contains the analysis to generate figure 3 of the review. In order to compare Hbt.sal and S. cer dynamics, we normalized each time series to the time 0, unshocked sample. This was similar to the normalization scheme conducted by Gasch et al. (2000). Supplemental figure 2 was also generated, showing that the Hbt. dynamics are not specific to a single condition.

##Set up:
```{r, warning=F, message=F}
#libraries
library(tidyverse); library(viridis); library(cowplot); library(grid); library(gridExtra)

#EXPRESSION DATA
scer <- read_csv("../output/cleaned_data/gasch_total.csv")
  scer[scer == 0] <- NA
  scer[scer == "#VALUE!"] <- NA
  scer <- scer[colSums(!is.na(scer)) > 0]
  scer[-1] <- lapply(scer[-1], function(x) as.numeric(as.character(x)))
sc.clust <- read_csv("../output/clustered_genelists/yeast2clusters.csv")
sc.clust$gene <-  str_replace_all(sc.clust$gene, "\\.", "-")

hbt <- read_csv("../output/cleaned_data/hbt_1000_var.csv")
hbt.clust <- read_csv("../output/clustered_genelists/hbt2clusters.csv")

#functions
source("00_custom_functions.R")
```

##YEAST
Use cluster information to filter expression data. Calculate the mean of all gene expression across conditions (as in analysis for figure 1):
```{r}
sc.esr <- scer[scer$gene %in% sc.clust$gene,]

t.yeast00 <- data.t(sc.esr, sc.esr$gene, scaled = "NULL")
  colnames(t.yeast00) <-  colnames(t.yeast00) %>% str_replace_all(., "\\.", "-")
yeast00.long <- gather(t.yeast00, gene, expression, -cond)
yeast00.long[yeast00.long == 0] <- NA

ex.yeast <- left_join(yeast00.long, sc.clust, by = "gene")
ex.yeast[is.na(ex.yeast)] <- 0

#get mean expression by conditon for clusters and overall
#detach(package:plyr)
ex.yeast %>%
  group_by(cluster, cond) %>%
  summarise(rep = length(expression), avg_expression = mean(expression),
            CI95 = conf_int95(expression)) -> y.avg.ex
```

####Severe versus mild stress
Dynamics of the two clusters when exposed to high levels of heat shock (25C to 37C shift) versus moderate heat shock (29C to 33C).
```{r, warning=FALSE, fig.height=6, fig.width= 6}
y.avg.ex[str_detect(y.avg.ex$cond, "Heat.*hs-1"),] %>%
  separate(cond, c("cond", "time"), "Shock ")  ->a
a$time <- as.integer(str_replace_all(a$time, ".in.*", ""))
a$cluster <- as.character(a$cluster)
a$cond <- rep("25C to 37C", length(a$cond))

y.avg.ex[str_detect(y.avg.ex$cond, "29C to"),] %>%
  separate(cond, c("cond", "time"), " - ")  ->b
b$time <- as.integer(str_replace_all(b$time, ".in.*", ""))
b$cluster <- as.character(b$cluster)

zeros <- data.frame("cluster" = c("1","2","1","2"),
                    "cond" = c("29C to 33C", "29C to 33C", "25C to 37C", "25C to 37C"),
                    "time" = c(0,0,0,0),
                    "rep" = rep(NA, 4),
                    "avg_expression" = rep(0, 4),
                    "CI95" = rep(NA, 4), stringsAsFactors = F)

#add a zero time point (lost when we normalized)
a <- bind_rows(a, zeros)

rbind(a,b) %>%
  ggplot(aes(x=time, y=avg_expression, shape = cond, color = cluster, group= interaction(cond, cluster)))+
  geom_point(show.legend = F, size = 3, fill = NA)+
  geom_line(aes(linetype = cond), size = 1.5, alpha = 0.8) +
  scale_color_viridis(option = "C", begin = 0.05, end = 0.65, discrete = T, guide = F) +
  scale_linetype_manual(labels = c("High", "Low"), values = c("25C to 37C" = 1, "29C to 33C" = 6)) +
  geom_hline(yintercept =0, color = "black") +
  xlim(0, 65)+ylim(-3,3)+
  labs(y = "Expression fold-change relative to unstressed baseline", x= "Time (minutes)", linetype = "Heat Shock") +
  theme_classic() + theme(axis.title.x=element_blank(), 
                          axis.title.y=element_blank(),
                          legend.title = element_text(family = "sans", size = 12),
                          legend.text = element_text(size = 10),
                          legend.justification=c(1,0), 
                          legend.position=c(.9, 0.85),  
                          legend.background = element_blank(),
                          legend.key = element_blank()) -> panel.A
panel.A


rbind(a,b) %>%
  ggplot(aes(x=time, y=avg_expression, shape = cond, color = cluster, group= interaction(cond, cluster)))+
  #geom_point(show.legend = F, size = 3, fill = NA)+
  #geom_line(aes(linetype = cond), size = 1.5, alpha = 0.8) +
  stat_smooth(aes(x=time, y=avg_expression, linetype = cond), method = lm, formula = y ~ poly(x, 3), se = FALSE, size =2) +
  scale_color_viridis(option = "C", begin = 0.05, end = 0.65, discrete = T, guide = F) +
  scale_linetype_manual(labels = c("High", "Low"), values = c("25C to 37C" = 1, "29C to 33C" = 3)) +
  geom_hline(yintercept =0, color = "black", size = 1) +
  xlim(0, 60)+ylim(-2.2,2.2)+
  labs(y = "Expression relative to unstressed baseline", x= "Time", linetype = "Stress Level") +
  theme_classic() + theme(axis.line = element_line(colour = "black", size = 1, linetype = "solid"),
                          axis.ticks = element_blank(), axis.text.x = element_blank(), axis.text.y = element_blank(),
                          axis.title = element_text(family = "sans", size = 16, face = "bold"),
                          legend.title = element_text(family = "sans", size = 14),
                          legend.text = element_text(size = 12, color = "black"),
                          legend.justification=c(1,0), 
                          legend.position=c(.9, 0.85),  
                          legend.background = element_blank(),
                          legend.key = element_blank()) -> graphical.abs
ggsave("../output/figures/graphical.png")
```

####Exposure versus removal from stress
Repeat analysis on exposure (25 to 37C) versus removal (37 to 25C) of from heat shock
```{r, echo=F, warning=FALSE, fig.height=6, fig.width= 6}
y.avg.ex[str_detect(y.avg.ex$cond, "Heat.*hs-1"),] %>%
  separate(cond, c("cond", "time"), "Shock ") -> a
a$time <- as.integer(str_replace_all(a$time, ".in.*", ""))
a$cluster <- as.character(a$cluster)
a$cond <- rep("25C to 37C", length(a$cond))

y.avg.ex[str_detect(y.avg.ex$cond, "37C "),] %>%
  separate(cond, c("cond", "time"), " shock - ") -> b
  b$time <- as.integer(str_replace_all(b$time, ".in.*", ""))
  b$cluster <- as.character(b$cluster)
 
zeros <- data.frame("cluster" = c("1","2","1","2"),
                    "cond" = c("37C to 25C", "37C to 25C", "25C to 37C", "25C to 37C"),
                    "time" = c(0,0,0,0),
                    "rep" = rep(NA, 4),
                    "avg_expression" = rep(0, 4),
                    "CI95" = rep(NA, 4), stringsAsFactors = F)
a <- bind_rows(a, zeros)

rbind(a,b) %>%
  ggplot(aes(x=time, y=avg_expression, shape = cond, color = cluster, group= interaction(cond, cluster)))+
  geom_point(show.legend = F, size = 3, fill = NA)+
  geom_line(aes(linetype = cond), size = 1.5, alpha = 0.8) +
  scale_color_viridis(option = "C", begin = 0.05, end = 0.65, discrete = T, guide = F) +
   scale_linetype_manual(labels = c("Constant", "Recovery"), values = c("25C to 37C" = 1, "37C to 25C" = 6)) +
  geom_hline(yintercept =0, color = "black") +
  xlim(0, 65)+ ylim(-3,3)+
  labs(y = "Expression fold-change relative to unstressed baseline", x= "Time (minutes)", linetype = "Stress Treatment") +
  theme_classic() + theme(axis.title.x=element_blank(), 
                          axis.title.y=element_blank(), 
                          legend.title = element_text(family = "sans", size = 12),
                          legend.text = element_text(size = 10),
                          legend.justification=c(1,0), 
                          legend.position=c(.9, 0.85),  
                          legend.background = element_blank(),
                          legend.key = element_blank())-> panel.B

panel.B
```

##HBT
Use cluster information to filter expression data. Calculate the mean of all gene expression across conditions (as in analysis for figure 1):
```{r, message=F}
t.hbt <- data.t(hbt, hbt$gene, scaled = "gene")
hbt.long <- gather(t.hbt, gene, expression, -cond)
hbt.long[hbt.long == 0] <- NA

ex.hbt <- left_join(hbt.long, hbt.clust)
ex.hbt[is.na(ex.hbt)] <- 0

#get mean expression by conditon for clusters and overall
ex.hbt %>%
  group_by(cluster, cond) %>%
  summarise(rep = length(expression), avg_expression = mean(expression),
            Std.error = sd(expression)/sqrt(length(expression))) -> hbt.avg.ex
```

Filter expression data by treatments with high and low paraquat:
```{r}
hbt.avg.ex[str_detect(hbt.avg.ex$cond, "PQ_Con.*0.25"),] %>%
  separate(cond, c("condition", "time"), "mM_") -> c
c$time <- as.integer(str_replace_all(c$time, "m.*", ""))
c$cluster <- as.character(c$cluster)

hbt.avg.ex[str_detect(hbt.avg.ex$cond, "PQ_Con.*1_4m"),] %>%
  separate(cond, c("condition", "time"), "mM_") -> d
d$time <- as.integer(str_replace_all(d$time, "min.*", ""))
#d$time <- as.integer(str_replace_all(c$time, "-.*", "0"))
d$cluster <- as.character(d$cluster)
```

####Severe versus mild stress
Normalize the high and low paraquat treatments to the time zero samples:
```{r, warning=FALSE, fig.height=6, fig.width= 6}
df2 <- 
  filter(c, time == 0) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))

full_join(c,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> c

df2 <- 
  filter(d, time == 0) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))
 
full_join(d,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> d

#zero transform (because a/a = 1, and we would like it to equal 0)
c %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression -1) -> c.1
c %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression +1) -> c.2
rbind(c.1, c.2) ->c

d %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression -1) -> d.1
d %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression +1) -> d.2
rbind(d.1, d.2) ->d

rbind(d,c) %>%
  ggplot(aes(x=time, y=avg_exp, shape = condition, color = cluster, group= interaction(condition, cluster)))+
  geom_point(show.legend = F, size = 3, fill = NA)+
  scale_shape_manual(values = c(17, 19))+
  geom_line(aes(linetype = condition), size = 1.5, alpha = 0.8) +
  scale_color_viridis(option = "C", begin = 0.05, end = 0.65, discrete = T, guide = F) + 
  scale_linetype_manual(labels = c("Low", "High"), 
                        values = c("PQ_Const._set_1_4" = 1, "PQ_Const._set_1_0.25" = 6)) +
  geom_hline(yintercept =0, color = "black") +
  xlim(c(0,250))+ylim(c(-8,8))+
  labs(y = "Expression fold-change relative to unstressed baseline", 
       x ="Time (minutes)", 
       linetype = "Oxidative Stress") +
  theme_classic() + theme(axis.title.x=element_blank(), 
                          axis.title.y=element_blank(),
                          legend.title = element_text(family = "sans", size = 12),
                          legend.text = element_text(size = 10),
                          legend.justification=c(1,0), 
                          legend.position=c(1, 0.90),  
                          legend.background = element_blank(),
                          legend.key = element_blank())-> panel.C
panel.C

```

####Exposure versus removal from stress
Repeat, but for samples under constant treatment of paraquat, or recovery from paraquat treatment  
```{r, echo=F, fig.height=6, fig.width= 6}
hbt.avg.ex[str_detect(hbt.avg.ex$cond, "PQ_C.*1_4mM"),] %>%
  separate(cond, c("condition", "time"), "mM_") -> c
c$time <- as.integer(str_replace_all(c$time, "m.*", ""))
c$cluster <- as.character(c$cluster)

hbt.avg.ex[str_detect(hbt.avg.ex$cond, "PQ_R.*2_4mM"),] %>%
  separate(cond, c("condition", "time"), "mM_") -> d
d$time <- as.integer(str_replace_all(d$time, "min.*", ""))
#d$time <- as.integer(str_replace_all(c$time, "-.*", "0"))
d$cluster <- as.character(d$cluster)

#normalize to the absolute value of the avg expression for time 0
df2 <- 
  filter(c, time == 0) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))

full_join(c,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> c

df2 <- 
  filter(d, time == 0) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))
 
full_join(d,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> d

#zero transform (because a/a = 1, and we would like it to equal 0)
c %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression -1) -> c.1
c %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression +1) -> c.2
rbind(c.1, c.2) -> c

d %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression +1) -> d.1
d %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression -1) -> d.2
rbind(d.1, d.2) ->d

rbind(c,d) %>%
  ggplot(aes(x=time, y=avg_exp, shape = condition, color = cluster, group= interaction(condition, cluster)))+
  geom_point(show.legend = F, size = 3, fill = NA)+
  geom_line(aes(linetype = condition), size = 1.5, alpha = 0.8) +
  scale_color_viridis(option = "C", begin = 0.05, end = 0.65, discrete = T, guide = F) + 
  scale_linetype_manual(labels = c("Constant", "Recovery"), 
                        values = c("PQ_Const._set_1_4" = 1, "PQ_Recov._set_2_4" = 6)) +
  geom_hline(yintercept =0, color = "black") +
  xlim(c(0,250))+ ylim(c(-8,8))+
  labs(y = "Expression fold-change relative to unstressed baseline", 
       x= "Time (minutes)", 
       linetype = "Stress Treatment") +
  theme_classic() + theme(axis.title.x=element_blank(), 
                          axis.title.y=element_blank(),
                          legend.title = element_text(family = "sans", size = 12),
                          legend.text = element_text(size = 10),
                          legend.justification=c(1,0), 
                          legend.position=c(1, 0.9),  
                          legend.background = element_blank(),
                          legend.key = element_blank()) -> panel.D
panel.D
```

##Bring all the panels together:
```{r, fig.height=10, fig.width=10, collapse=T, warning=F}
plot_grid(panel.A, panel.B) -> plots
title1 <- ggdraw() + 
  draw_label("S. cerevisiae", fontfamily = "sans", size = 14,
             fontface = 'bold.italic')

plot_grid(panel.C, panel.D) -> plots2
title2 <- ggdraw() + 
  draw_label("Hbt. salinarum", fontfamily = "sans", size = 14,
             fontface = 'bold.italic', vjust = 1)

plot_grid(title1, plots, title2, plots2,ncol = 1, rel_heights = c(0.2, 1, 0.35, 1)) -> plots

plots <- plots + draw_plot_label(c("(a)", "(b)", "(c)", "(d)"), c(0, 0.5, 0, 0.5), c(0.95,0.95,0.43,0.43))  


y.grob <- textGrob("Expression fold-change relative to unstressed baseline", 
                   gp=gpar(fontface="bold", col="black", fontsize=12), rot=90)
x.grob <- textGrob("time (minutes)", 
                   gp=gpar(fontface="bold", col="black", fontsize=12))
plot(plots)
```

###export figure
```{r, results="hide"}
#save plot
tiff("../output/figures/Figure3.tif", units="in", width=5, height=5, res=300)
grid.arrange(arrangeGrob(plots, left = y.grob, bottom = x.grob))
dev.off()

pdf("../output/figures/Figure3.pdf", width = 10, height = 10)
grid.arrange(arrangeGrob(plots, left = y.grob, bottom = x.grob))
dev.off()
```

##Same dynamics are present in other conditions (supplemental fig2)
H2o2 recovery vs. constant:
```{r, include=FALSE}
hbt.avg.ex[str_detect(hbt.avg.ex$cond, "H2O2_C.*_2_25mM"),] %>%
  separate(cond, c("condition", "time"), "mM_") -> c
c$time <- as.integer(str_replace_all(c$time, "m.*", ""))
c$cluster <- as.character(c$cluster)

hbt.avg.ex[str_detect(hbt.avg.ex$cond, "H2O2_R.*_3_25mM"),] %>%
  separate(cond, c("condition", "time"), "mM_") -> d
d$time <- as.integer(str_replace_all(d$time, "min.*", ""))
#d$time <- as.integer(str_replace_all(c$time, "-.*", "0"))
d$cluster <- as.character(d$cluster)

#normalize to the absolute value of the avg expression for time 0
df2 <- 
  filter(c, time == 0) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))

full_join(c,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> c

df2 <- 
  filter(d, time == 0) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))
 
full_join(d,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> d

#zero transform (because a/a = 1, and we would like it to equal 0)
c %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression +1) -> c.1
c %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression -1) -> c.2
rbind(c.1, c.2) -> c

d %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression +1) -> d.1
d %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression -1) -> d.2
rbind(d.1, d.2) ->d

rbind(c,d) %>%
  ggplot(aes(x=time, y=avg_exp, shape = condition, color = cluster, group= interaction(condition, cluster)))+
  geom_point(show.legend = F, size = 3, fill = NA)+
  geom_line(aes(linetype = condition), size = 1.5, alpha = 0.8) +
  scale_color_viridis(option = "C", begin = 0.05, end = 0.65, discrete = T, guide = F) + 
  scale_linetype_manual(labels = c("H2O2 Constant", "H2O2 Recovery"), 
                        values = c("H2O2_Const._set_2_25" = 1, "H2O2_Recov._set_3_25" = 6)) +
  geom_hline(yintercept =0, color = "black") +
  xlim(c(0,250))+ ylim(c(-5,15))+
  labs(y = "Expression fold-change relative to unstressed baseline", 
       x= "Time (minutes)", 
       linetype = "Stress Treatment", 
       title =  "25mM Hydrogen Peroxide") +
  theme_classic() + theme(axis.title.x=element_blank(), 
                          axis.title.y=element_blank(),
                          legend.title = element_text(family = "sans", size = 12),
                          legend.text = element_text(size = 10),
                          legend.justification=c("right", "top"), 
                          legend.position=c(.95, 0.95),  
                          legend.background = element_blank(),
                          legend.key = element_blank(),
                          plot.title = element_text(hjust = 0.5, size = 14)) -> h2o2
```

copper doses:
```{r, include=FALSE}
hbt.avg.ex[str_detect(hbt.avg.ex$cond, "sDura3_pCu.*100"),] %>%
  separate(cond, c("condition", "time"), "_t") -> c
c$time <- as.integer(str_replace_all(c$time, "m.*", ""))
c$cluster <- as.character(c$cluster)

hbt.avg.ex[str_detect(hbt.avg.ex$cond, "sDura3_pCu.*700"),] %>%
  separate(cond, c("condition", "time"), "_t") -> d
d$time <- as.integer(str_replace_all(d$time, "m.*", ""))
d$cluster <- as.character(d$cluster)

#normalize to the absolute value of the avg expression for time 0
df2 <- 
  filter(c, time == -15) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))

full_join(c,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> c

df2 <- 
  filter(d, time == -15) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))
 
full_join(d,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> d

#zero transform (because a/a = 1, and we would like it to equal 0)
c %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression-1) -> c.1
c %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression+1) -> c.2
rbind(c.1, c.2) -> c

d %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression -1) -> d.1
d %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression -1) -> d.2
rbind(d.1, d.2) ->d

zeros <- data.frame("cluster" = c("1","2","1","2"),
                    "condition" = c("sDura3_pCu_d0.100mM", "sDura3_pCu_d0.100mM", "sDura3_pCu_d0.700mM", "sDura3_pCu_d0.700mM"),
                    "time" = c(0,0,0,0),
                    "rep" = rep(NA, 4),
                    "avg_exp" = c(0,0,0,0),
                    "CI95" = rep(NA, 4), stringsAsFactors = F)
c <- bind_rows(c, zeros)

rbind(c,d) %>%
  ggplot(aes(x=time, y=avg_exp, shape = condition, color = cluster, group= interaction(condition, cluster)))+
  geom_point(show.legend = F, size = 3, fill = NA)+
  geom_line(aes(linetype = condition), size = 1.5, alpha = 0.8) +
  scale_color_viridis(option = "C", begin = 0.05, end = 0.65, discrete = T, guide = F) + 
  scale_linetype_manual(labels = c("0.1mM", "0.7mM"), 
                        values = c("sDura3_pCu_d0.700mM" = 1, "sDura3_pCu_d0.100mM" = 6)) +
  geom_hline(yintercept =0, color = "black") +
  xlim(c(0,250))+ ylim(c(-5,10))+
  labs(y = "Expression fold-change relative to unstressed baseline", 
       x= "Time (minutes)", 
       linetype = "Stress", 
      title =  "Copper") +
  theme_classic() + theme(axis.title.x=element_blank(), 
                          axis.title.y=element_blank(),
                          legend.title = element_text(family = "sans", size = 12),
                          legend.text = element_text(size = 10),
                          legend.justification=c("right", "top"), 
                          legend.position=c(.95, 0.95),  
                          legend.background = element_blank(),
                          legend.key = element_blank(),
                          plot.title = element_text(hjust = 0.5, size = 14)) -> copper
```

zinc doses:
```{r, include=FALSE}
hbt.avg.ex[str_detect(hbt.avg.ex$cond, "sDura3_pZn.*0.00"),] %>%
  separate(cond, c("condition", "time"), "_t") -> c
c$time <- as.integer(str_replace_all(c$time, "m.*", ""))
c$cluster <- as.character(c$cluster)


hbt.avg.ex[str_detect(hbt.avg.ex$cond, "sDura3_pZn.*0.02"),] %>%
  separate(cond, c("condition", "time"), "_t") -> d
d$time <- as.integer(str_replace_all(d$time, "m.*", ""))
d$cluster <- as.character(d$cluster)

hbt.avg.ex[str_detect(hbt.avg.ex$cond, "sDura3_pZn.*0.05"),] %>%
  separate(cond, c("condition", "time"), "_t") -> e
e$time <- as.integer(str_replace_all(e$time, "m.*", ""))
e$cluster <- as.character(e$cluster)

#normalize to the absolute value of the avg expression for time 0
df2 <- 
  filter(c, time == 15) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))

full_join(c,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> c

df2 <- 
  filter(d, time == 15) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))
 
full_join(d,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> d

df2 <- 
  filter(e, time == 15) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))
 
full_join(e,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> e

#zero transform (because a/a = 1, and we would like it to equal 0)
c %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression+1) -> c.1
c %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression-1) -> c.2
rbind(c.1, c.2) -> c

d %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression -1) -> d.1
d %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression +1) -> d.2
rbind(d.1, d.2) ->d

e %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression -1) -> e.1
e %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression +1) -> e.2
rbind(e.1, e.2) ->e

rbind(c,d,e) %>%
  ggplot(aes(x=time, y=avg_exp, shape = condition, color = cluster, group= interaction(condition, cluster)))+
  geom_point(show.legend = F, size = 3, fill = NA)+
  geom_line(aes(linetype = condition), size = 1.5, alpha = 0.9) +
  scale_color_viridis(option = "C",begin = 0.05, end = 0.65, discrete = T, guide = F) + 
  scale_linetype_manual(labels = c("0.005mM", "0.02mM", "0.05mM"), 
                        values = c("sDura3_pZn_d0.020mM" = 6, "sDura3_pZn_d0.005mM" = 3, "sDura3_pZn_d0.050mM" = 1)) +
  geom_hline(yintercept =0, color = "black") +
  xlim(c(0,200))+ ylim(c(-8,8))+
  labs(y = "Expression fold-change relative to unstressed baseline", 
       x= "Time (minutes)", 
       linetype = "Stress", 
      title =  "Zinc") +
  theme_classic() + theme(axis.title.x=element_blank(), 
                          axis.title.y=element_blank(),
                          legend.title = element_text(family = "sans", size = 12),
                          legend.text = element_text(size = 10),
                          legend.justification=c("right", "top"), 
                          legend.position=c(.95, 0.95),  
                          legend.background = element_blank(),
                          legend.key = element_blank(),
                          plot.title = element_text(hjust = 0.5, size = 14)) -> zinc
```

Temp+circadian:
```{r, include=FALSE}
hbt.avg.ex[str_detect(hbt.avg.ex$cond, "42C"),] %>%
  separate(cond, c("condition", "time"), "_42C_") -> c
c$condition <- str_replace_all(c$condition, "125w/m2", "42C")
c$time <- as.integer(str_replace_all(c$time, "min.*", ""))
c$time <- c$time - 1440
c$cluster <- as.character(c$cluster)

hbt.avg.ex[str_detect(hbt.avg.ex$cond, "30C"),] %>%
  separate(cond, c("condition", "time"), "_30C_") -> d
d$condition <- str_replace_all(d$condition, "ling", "_30C")
d$time <- as.integer(str_replace_all(d$time, "min.*", ""))
d$time <- d$time - 1440
d$cluster <- as.character(d$cluster)

#normalize to the absolute value of the avg expression for time 0
df2 <- 
  filter(c, time == 0) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))

full_join(c,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> c

df2 <- 
  filter(d, time == 0) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))
 
full_join(d,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> d

#zero transform (because a/a = 1, and we would like it to equal 0)
c %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression+1) -> c.1
c %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression-1) -> c.2
rbind(c.1, c.2) -> c

d %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression +1) -> d.1
d %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression -1) -> d.2
rbind(d.1, d.2) ->d

rbind(c,d) %>%
  ggplot(aes(x=time, y=avg_exp, shape = condition, color = cluster, group= interaction(condition, cluster)))+
  geom_point(show.legend = F, size = 3, fill = NA)+
  geom_line(aes(linetype = condition), size = 1.5, alpha = 0.8) +
  scale_color_viridis(option = "C", begin = 0.05, end = 0.65, discrete = T, guide = F) + 
  scale_linetype_manual(labels = c("42C", "30C"), 
                        values = c("circadian__dark_cycling_42C" = 1, "circadian__dark5_cyc_30C" = 6)) +
  geom_hline(yintercept =0, color = "black") +
  xlim(c(0,1500))+ ylim(c(-1,2))+
  labs(y = "Expression fold-change relative to unstressed baseline", 
       x= "Time (minutes)", 
       linetype = "Stress", 
      title =  "Temperature + Circadian cycling") +
  theme_classic() + theme(axis.title.x=element_blank(), 
                          axis.title.y=element_blank(),
                          legend.title = element_text(family = "sans", size = 12),
                          legend.text = element_text(size = 10),
                          legend.text.align = 1,
                          legend.justification=c("right", "top"), 
                          legend.position=c(.95, 0.95),  
                          legend.background = element_blank(),
                          legend.key = element_blank(),
                          plot.title = element_text(hjust = 0.5, size = 14)) -> temp
```

Outliers:
UV:
```{r, include=FALSE}
hbt.avg.ex[str_detect(hbt.avg.ex$cond, "gamma4__0"),] %>%
  separate(cond, c("condition", "time"), "-") -> c
c$time <- as.integer(str_replace_all(c$time, "min", ""))
c$cluster <- as.character(c$cluster)

hbt.avg.ex[str_detect(hbt.avg.ex$cond, "gamma__2"),] %>%
  separate(cond, c("condition", "time"), "-") -> d
d$time <- as.integer(str_replace_all(d$time, "m", ""))
d$cluster <- as.character(d$cluster)

#normalize to the absolute value of the avg expression for time 0
df2 <- 
  filter(c, time == 0) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))

full_join(c,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> c

df2 <- 
  filter(d, time == 0) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))
 
full_join(d,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> d

#zero transform (because a/a = 1, and we would like it to equal 0)
c %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression-1) -> c.1
c %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression+1) -> c.2
rbind(c.1, c.2) -> c

d %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression -1) -> d.1
d %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression +1) -> d.2
rbind(d.1, d.2) ->d

rbind(c,d) %>%
  ggplot(aes(x=time, y=avg_exp, shape = condition, color = cluster, group= interaction(condition, cluster)))+
  geom_point(show.legend = F, size = 3, fill = NA)+
  geom_line(aes(linetype = condition), size = 1.5, alpha = 0.8) +
  scale_color_viridis(option = "C",begin = 0.05, end = 0.65, discrete = T, guide = F) + 
  scale_linetype_manual(labels = c("2500 Gy", "0 Gy"), 
                        values = c("gamma__2500gy" = 1, "gamma4__0000gy" = 6)) +
  geom_hline(yintercept =0, color = "black") +
  xlim(c(0,250))+ ylim(c(-2,5))+
  labs(y = "Expression fold-change relative to unstressed baseline", 
       x= "Time (minutes)", 
       linetype = "Stress", 
      title =  "UV (gamma) radiation") +
  theme_classic() + theme(axis.title.x=element_blank(), 
                          axis.title.y=element_blank(),
                          legend.title = element_text(family = "sans", size = 12),
                          legend.text = element_text(size = 10),
                          legend.justification=c("right", "top"), 
                          legend.position=c(.95, 0.95),  
                          legend.background = element_blank(),
                          legend.key = element_blank(),
                          plot.title = element_text(hjust = 0.5, size = 14)) -> UV.rad
```

EMS (UV + DNA damage)
```{r,include=FALSE}
hbt.avg.ex[str_detect(hbt.avg.ex$cond, "EMS.*__1"),] %>%
  separate(cond, c("condition", "time"), "-") -> c
c$time <- as.integer(c$time)
c$cluster <- as.character(c$cluster)

hbt.avg.ex[str_detect(hbt.avg.ex$cond, "EMS.*__0"),] %>%
  separate(cond, c("condition", "time"), "-") -> d
d$time <- as.integer(d$time)
d$cluster <- as.character(d$cluster)

#normalize to the absolute value of the avg expression for time 0
df2 <- 
  filter(c, time == 0) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))

full_join(c,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> c

df2 <- 
  filter(d, time == 0) %>% 
  select(c("cluster", "avg_expression")) %>%
  setNames(.,c("cluster", "a"))
 
full_join(d,df2,by="cluster") %>%
   group_by(cluster) %>%
   mutate(avg_expression=avg_expression/abs(a)) %>%
   select(-a) -> d

#zero transform (because a/a = 1, and we would like it to equal 0)
c %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression-1) -> c.1
c %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression+1) -> c.2
rbind(c.1, c.2) -> c

d %>% filter(cluster==1) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression -1) -> d.1
d %>% filter(cluster==2) %>%
  group_by(cluster, condition, time) %>%
  summarise(avg_exp = avg_expression +1) -> d.2
rbind(d.1, d.2) ->d

rbind(c,d) %>%
  ggplot(aes(x=time, y=avg_exp, shape = condition, color = cluster, group= interaction(condition, cluster)))+
  geom_point(show.legend = F, size = 3, fill = NA)+
  geom_line(aes(linetype = condition), size = 1.5, alpha = 0.8) +
  scale_color_viridis(option = "C", begin = 0.05, end = 0.65, discrete = T, guide = F) + 
  scale_linetype_manual(labels = c("0mM EMS", " 100mM EMS"), 
                        values = c("EMS_Treated__100000uM" = 1, "EMS_Treated__000uM" = 6)) +
  geom_hline(yintercept =0, color = "black") +
  xlim(c(0,120))+ ylim(c(-2,5))+
  labs(y = "Expression fold-change relative to unstressed baseline", 
       x= "Time (minutes)", 
       linetype = "Stress", 
      title =  "EMS") +
  theme_classic() + theme(axis.title.x=element_blank(), 
                          axis.title.y=element_blank(),
                          legend.title = element_text(family = "sans", size = 12),
                          legend.text = element_text(size = 10),
                          legend.justification=c("right", "top"), 
                          legend.position=c(.95, 0.95),  
                          legend.background = element_blank(),
                          legend.key = element_blank(),
                          plot.title = element_text(hjust = 0.5, size = 14)) -> EMS
```

Combine:
```{r, fig.height=10, fig.width=6, warning=F, message=F}
title1 <-  ggdraw()+
  draw_label("Exposure vs. Recovery",  fontfamily = "sans", size = 14, fontface = 'bold', hjust = 0)
title2 <-  ggdraw()+
  draw_label("Dose sensitivity",  fontfamily = "sans", size = 14, fontface = 'bold', hjust = 0)
title3 <-  ggdraw()+
  draw_label("DNA and UV damage may elicite a distinct response",  fontfamily = "sans", size = 14, fontface = 'bold', hjust = 0)

plot_grid(title1, h2o2, title2, zinc,
  #plot_grid(title3, NULL, nrow = 1), plot_grid(UV.rad, EMS, NULL, nrow = 1),
  ncol = 1, rel_heights = c(0.1, 1, 0.1, 1), labels = c("(a)", NA, "(b)", NA)) -> Supp2
Supp2
```

export:
```{r, warning=F,message=F}
pdf("../output/figures/SuppFig2.pdf", width = 6, height = 10)
grid.arrange(arrangeGrob(Supp2, left = y.grob, bottom = x.grob))
#plot(Supp2)
dev.off()

#save plot
```
